+++ .env
@@ -1,0 +1,3 @@
+SECRET_KEY='9-8b"dts-zMEH!bt_(RN-mx<nf4/P>;k!KZRcM8Svd`EA-~!au'
+DEBUG=True
+CLOUDINARY_URL='cloudinary://365149543497821:tY62gbhFmp0TrMmaCThsTuIDSSc@hm65a3i1o'
