+++ Procfile.windows
@@ -1,0 +1,2 @@
+release: python manage.py migrate
+web: python manage.py runserver 0.0.0.0:5000
