+++ core/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from core.views import products, home, product_details, add_to_cart, empty_cart, shopping_cart, checkout, remove_shopping_cart_item, contact, wishlist, add_to_wishlist, remove_wishlist_item
+from core.views import products, home, product_details, add_to_cart, empty_cart, shopping_cart, checkout, remove_shopping_cart_item, contact 
 
 urlpatterns = [
     path("products/", products, name="products"),
@@ -15,8 +15,4 @@ urlpatterns = [
     path("contact/", contact, name="contact"),
     path("shopping-cart/remove-item/<session_item_id>",
          remove_shopping_cart_item, name="remove_shopping_cart_item"),
-    path("wishlist/", wishlist, name="wishlist"),
-    path("add-to-wishlist/", add_to_wishlist, name="add_to_wishlist"),
-    path("wishlist/remove-item/<session_item_id>",
-         remove_wishlist_item, name="remove_wishlist_item"),
 ]
