+++ core/views.py
@@ -168,78 +168,3 @@ def checkout(request):
 def contact(request):    
     return render(request, "contact.html")
 
-def wishlist(request):
-    context = wishlist_context(request)
-    return TemplateResponse(request, "wishlist.html", context)
-
-
-def remove_wishlist_item(request, session_item_id):
-    """Removes an item from the session wishlist, by its unique ID in the session."""
-
-    # Remove the item from member db wishlist storage.
-    request.user.member.cartitem_set.filter(
-        session_item_id=session_item_id).delete()
-
-    # If the cart is empty, just redirect to the home page.
-    if "wishlist" not in request.session:
-        return redirect("/")
-
-    new_items = []
-    for item in request.session["wishlist"]:
-        if item["session_item_id"] != session_item_id:
-            new_items.append(item)
-
-    request.session["wishlist"] = new_items
-
-    return redirect("/wishlist")
-
-
-def add_to_wishlist(request):
-    if "wishlist" not in request.session:
-        wishlist = []
-    else:
-        wishlist = request.session["wishlist"]
-
-    session_item_id = str(uuid.uuid4())
-
-    if request.user.is_authenticated:
-        wi = WishlistItem(
-            member_id=request.user.member.id,
-            item_id=request.POST["product_id"],
-            session_item_id=session_item_id,
-            quantity=request.POST["quantity"]
-        )
-        wi.color = Color.objects.get(hex_value=request.POST["color"])
-        wi.size = Size.objects.get(name=request.POST["size"])
-        wi.save()
-
-    wishlist.append({
-        "session_item_id": session_item_id,
-        "product_id": request.POST["product_id"],
-        "color": request.POST["color"],
-        "size": request.POST["size"],
-        "quantity": request.POST["quantity"],
-    })
-
-    request.session["wishlist"] = wishlist
-
-    return redirect("/products")
-
-
-def wishlist_context(request):
-    """Creates a context suitable for wishlist and checkout pages, containing products, subtotal, 
-    shipping and total."""
-    context = {
-        "items": [],
-    }
-    if "wishlist" in request.session:
-        for item in request.session["wishlist"]:
-            product = Product.objects.get(id=item["product_id"])
-            context["items"].append({
-                "session_item_id": item["session_item_id"],
-                "product": product,
-                "quantity": item["quantity"],
-                "color": item["color"],
-                "size": item["size"],
-            })
-    return context
